% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.R
\name{cgv_camera}
\alias{cgv_camera}
\title{Set Camera Position and Direction}
\usage{
cgv_camera(viewer, position = c(0, 0, 5), target = c(0, 0, 0), up = c(0, 1, 0))
}
\arguments{
\item{viewer}{External pointer returned by \code{cgv_viewer}.}

\item{position}{Numeric vector of length 3 (x, y, z).}

\item{target}{Numeric vector of length 3 — look-at point.}

\item{up}{Numeric vector of length 3 — up direction.}
}
\value{
Invisible \code{NULL}.
}
\description{
Set Camera Position and Direction
}
