% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_render.R
\name{cgv_clear_path}
\alias{cgv_clear_path}
\title{Clear Path Highlight}
\usage{
cgv_clear_path(viewer)
}
\arguments{
\item{viewer}{External pointer returned by \code{cgv_viewer}.}
}
\value{
Invisible \code{NULL}.
}
\description{
Remove path highlight and restore original node colors and sizes.
}
