% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgvR-package.R
\docType{package}
\name{cgvR-package}
\alias{cgvR}
\alias{cgvR-package}
\title{cgvR: Interactive 3D Visualization of Large Cayley Graphs via Vulkan}
\description{
Interactive 3D visualization of large Cayley graphs using Vulkan-based rendering based on 'Datoviz' library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Zabis13/cgvR}
  \item Report bugs at \url{https://github.com/Zabis13/cgvR/issues}
}

}
\author{
\strong{Maintainer}: Yuri Baramykov \email{lbsbmsu@mail.ru}

Other contributors:
\itemize{
  \item Cyrille Rossant (Author of the Datoviz library) [contributor, copyright holder]
}

}
\keyword{internal}
